+++ QGscrap.py
