+++ main.py
@@ -1,2 +1,4 @@
-from trainee_generator import *
+from sparta_simulator import *
+
+
 sparta_simulator()
